+++ FExploradora.h
@@ -6,7 +6,7 @@ class FExploradora : public Formiga{
 public:
     FExploradora(int linha, int coluna, int id, Ninho* ninho_a_que_pertence);
     void consomeEnergia(int mov_efetivo) override;
-    int percentage_takes_from_migalha() const override;
+    float percentage_takes_from_migalha() const override;
