+++ Migalha.h
@@ -5,13 +5,13 @@
 
 class Migalha {
     Ponto local_m;
-    int energia_m, energia_inicial;
+    float energia_m, energia_inicial;
 public:
-    Migalha(int linha, int coluna, int energini);
+    Migalha(int linha, int coluna, float energini);
     Ponto getPonto() const;
     string getInfo()const;
-    int getEnergia()const;
-    int loseEnergy(int percentage_of_energy_to_take);
+    float getEnergia()const;
+    float loseEnergy(float percentage_of_energy_to_take);
     void iteracao();
     virtual ~Migalha();
 private:
