+++ Ninho.cpp
@@ -4,7 +4,7 @@
 
 int Ninho::n_ninhos = 0;
 
-Ninho::Ninho(int linha, int coluna, int energ_init_ninho, int def_p_novaformiga, int def_energ_iter): local_n(linha, coluna),
+Ninho::Ninho(int linha, int coluna, float energ_init_ninho, int def_p_novaformiga, int def_energ_iter): local_n(linha, coluna),
         energia_n(energ_init_ninho), p_novaformiga(def_p_novaformiga), energ_iter(def_energ_iter), id_n(++n_ninhos), energia_inicial(energ_init_ninho){}
 
 Ponto Ninho::getPonto() const{return local_n;}
@@ -17,7 +17,7 @@ string Ninho::getInfo() const{
 int Ninho::getId() const{
     return id_n;
 }
-bool Ninho::setenergia_n(int addenerg){
+bool Ninho::setenergia_n(float addenerg){
     if(energia_n + addenerg > 0){
         energia_n += addenerg;
         return true;
